# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LoveZone API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /states:
    get:
      operationId: listStates
      tags: [states]
      summary: List all states
      responses:
        "200":
          description: List of states
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/State"
    post:
      operationId: createState
      tags: [states]
      summary: Create a state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StateInput"
      responses:
        "201":
          description: Created state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/State"
  /states/{slug}:
    get:
      operationId: getState
      tags: [states]
      summary: Get state by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: State details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/State"
        "404":
          description: Not found
    patch:
      operationId: updateState
      tags: [states]
      summary: Update a state
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StateUpdate"
      responses:
        "200":
          description: Updated state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/State"
    delete:
      operationId: deleteState
      tags: [states]
      summary: Delete a state
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # --- Import ---
...
components:
  schemas:
...
    State:
      type: object
      required: [id, name, slug, listingCount]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        listingCount:
          type: integer
        createdAt:
          type: ["string", "null"]
    StateInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
    StateUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        metaTitle:
          type: string
        metaDescription:
          type: string
...